Move PathParser to separate module
