+++ components/parser/common/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    alias(libs.plugins.valkyrie.kmp)
+}
+
+kotlin {
+    sourceSets {
+        commonMain.dependencies {
+            implementation(projects.components.ir)
+        }
+    }
+}
