+++ components/parser/common/src/commonMain/kotlin/io/github/composegears/valkyrie/parser/common/PathNode.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.parser.jvm.xml
+package io.github.composegears.valkyrie.parser.common
 
 import io.github.composegears.valkyrie.ir.IrPathNode
 import io.github.composegears.valkyrie.ir.IrPathNode.ArcTo
