+++ components/parser/jvm/xml/build.gradle.kts
@@ -4,6 +4,6 @@ plugins {
 
 dependencies {
     implementation(projects.components.ir)
-
+    implementation(projects.components.parser.common)
     implementation(libs.xpp3)
 }
