+++ settings.gradle.kts
@@ -55,6 +55,7 @@ include("components:generator:iconpack")
 include("components:generator:imagevector")
 include("components:ir")
 include("components:ir-compose")
+include("components:parser:common")
 include("components:parser:jvm:svg")
 include("components:parser:jvm:xml")
 include("components:parser:ktfile")
